fix bugs

+++ container/CContainerInfo.cpp
@@ -255,15 +255,15 @@ bool CContainerInfo::save(CString file)
 		m_range.put_Item(COleVariant((long)i),COleVariant((long)2),COleVariant(str));
 		str.Format(_T("%.2f(%.2f)"), iter->conThick, iter->conCalcThick);
 		m_range.put_Item(COleVariant((long)i),COleVariant((long)3),COleVariant(str));
-		str.Format(_T("%.2f"), iter->conHight);
+		str.Format(_T("%d"), (int)iter->conHight);
 		m_range.put_Item(COleVariant((long)i),COleVariant((long)4),COleVariant(str));
 		str.Format(_T("%.2f(%.2f)"), iter->capThick, iter->capCalcThick);
 		m_range.put_Item(COleVariant((long)i),COleVariant((long)5),COleVariant(str));
-		str.Format(_T("%.2f"), iter->capHight);
+		str.Format(_T("%d"), (int)iter->capHight);
 		m_range.put_Item(COleVariant((long)i),COleVariant((long)6),COleVariant(str));
-		str.Format(_T("%.2f"), iter->totalHight);
+		str.Format(_T("%d"), (int)iter->totalHight);
 		m_range.put_Item(COleVariant((long)i),COleVariant((long)7),COleVariant(str));
-		str.Format(_T("%.2f"), iter->weight);
+		str.Format(_T("%d"), (int)iter->weight);
 		m_range.put_Item(COleVariant((long)i),COleVariant((long)8),COleVariant(str));
 
 		i++;
